+++ front-end/src/NavBar.js
@@ -3,7 +3,7 @@ import { Link } from "react-router-dom";
 import Navbar from 'react-bootstrap/Navbar';
 
 
-const NavBar = () => {
+const NavBar = (props) => {
   return (
     <Navbar bg = "light" sticky="top" >
         <Navbar.Collapse className="justify-content-center">
@@ -13,9 +13,14 @@ const NavBar = () => {
         </Navbar.Collapse> 
         <Navbar.Collapse className="justify-content-end">
           {!localStorage.getItem("token")? (
-              <Link className = "App-link" to="/signup">
+            <Link className = "App-link" to={{
+                pathname: '/signup',
+                state: {
+                  changeState: props.changeState
+                }
+            }}>
                   <Navbar.Brand > Sign Up </Navbar.Brand>
-              </Link>
+            </Link>
           ) : (
               <Link className = "App-link" to="/logout">
                   <Navbar.Brand > Log Out </Navbar.Brand>
